requirements.txt
python-consul
netifaces
